feature/VIH-7015 : Split out service tests
